Corrección, excepciones y buenas practicas en el backend

+++ demo/src/main/java/rookie/brain/clickerGame/Service/LevelService.java
@@ -1,9 +1,10 @@
 package rookie.brain.clickerGame.Service;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import rookie.brain.clickerGame.Entity.Level;
+import rookie.brain.clickerGame.Exception.LevelNotFoundException;
 import rookie.brain.clickerGame.Repository.LevelRepository;
+import rookie.brain.clickerGame.Utils.ExceptionConstants;
 
 import java.util.List;
 
@@ -25,10 +26,14 @@ public class LevelService {
     }
 
     public Level getLevelById(Long id) {
-        return levelRepository.findById(id).orElse(null);
+        return levelRepository.findById(id)
+                .orElseThrow(() -> new LevelNotFoundException(String.format(ExceptionConstants.LEVEL_NOT_FOUND, id)));
     }
 
     public void deleteLevel(Long id) {
+        if (!levelRepository.existsById(id)) {
+            throw new LevelNotFoundException(String.format(ExceptionConstants.LEVEL_NOT_FOUND, id));
+        }
         levelRepository.deleteById(id);
     }
 }
