+++ demo/src/main/java/rookie/brain/clickerGame/Service/RankingService.java
@@ -3,7 +3,9 @@ package rookie.brain.clickerGame.Service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import rookie.brain.clickerGame.Entity.Ranking;
+import rookie.brain.clickerGame.Exception.RankingNotFoundException;
 import rookie.brain.clickerGame.Repository.RankingRepository;
+import rookie.brain.clickerGame.Utils.ExceptionConstants;
 
 import java.util.List;
 
@@ -22,10 +24,16 @@ public class RankingService {
     }
 
     public Ranking getRankingById(Long id) {
-        return rankingRepository.findById(id).orElse(null);
+        return rankingRepository.findById(id)
+                .orElseThrow(() -> new RankingNotFoundException(
+                        String.format(ExceptionConstants.RANKING_NOT_FOUND, id)));
     }
 
     public void deleteRanking(Long id) {
+        if (!rankingRepository.existsById(id)) {
+            throw new RankingNotFoundException(
+                    String.format(ExceptionConstants.RANKING_NOT_FOUND, id));
+        }
         rankingRepository.deleteById(id);
     }
 }
